Allows use of hash of image id + workspace hash for caching and run/exec config
If the image changed, bazel would still use the old cache, which might be invalid if some of the
rules violate hermeticity, like is the case for python. As an alternative, we provide a flag that
mixes the image id into the workspace hash, effectivelly pretending to be a new workspace if the
image changes.

It also adds some flexibility for passing flags to exec or run, since they might be useful to define
different behaviors for the system.