Added interactive run option
When we do `bazel run //my:path` and the code is interactive, bazel will not allow you to
interact with it, just closing the stdin. This might become an issue specially for debuggers, since
we get an interactive session to the code in case of an exception.

The bazel solution to this is to use `--script_path` to write a script and then the user can execute
it. However, since we're running bazel from the container, that path might not be accessible from
the outside and the script should be run inside the container. To simplify this task for the user,
when the bazel command is `run`, we by default use this method of running through a script.