Added cleanup command
If an image isn't in the active list and has been around for more than the specified timescale, then
that image is removed. Note that this will not remove images still attached to containers, since
they can be used for debugging. Remember to remoe the containers first, but this will give you the
default docker warning that it couldn't remove as a reminder.